requirements.txt
altair
numpy
pandas
pydeck
streamlit
plotly.graph_objects
